python/agents/status_messenger/requirements.txt
fastapi
jinja2
uvicorn
requests
pydantic
google-cloud-aiplatform
google-api-core
google-genai
google-adk